update navbar with tailwind
