+++ pluralistic/src/components/Home.css
@@ -1,0 +1,7 @@
+.home{
+    width:100%;
+    height: 100vh;
+    background-color: linear-gradient(45deg,rgb(182, 13, 216)12%,
+    rgb(168, 12, 121)80%,
+    rgb(236, 223, 28)4%);
+}
