+++ pluralistic/src/components/Home.jsx
@@ -1,10 +1,13 @@
-import React from 'react'
-
+import React from 'react';
+import './Home.css';
 const Home = () => {
   return (
-    <div>
+    <>
+
+    <div className='home'>
         
     </div>
+    </>
   )
 }
